Create outdated-package-processor dir
