+++ src/core/OutdatedPackagesProcessor.test.ts
@@ -3,6 +3,7 @@ import {
   right,
   type Either
 } from 'fp-ts/lib/Either'
+import type { OutdatedPackageProcessor } from '../outdated-package-processor'
 import { DependencyType } from '../package-json'
 import {
   SemVer,
@@ -10,7 +11,6 @@ import {
 } from '../semver'
 import type { FailedResult } from './FailedResult'
 import type { OutdatedPackage } from './OutdatedPackage'
-import type { OutdatedPackageProcessor } from './OutdatedPackageProcessor'
 import { OutdatedPackagesProcessor } from './OutdatedPackagesProcessor'
 import type { SucceededResult } from './SucceededResult'
