+++ src/core/index.ts
@@ -1,8 +1,5 @@
 export * from './FailedResult'
 export * from './OutdatedPackage'
-export * from './OutdatedPackageProcessor'
-export * from './OutdatedPackageProcessorCreator'
 export * from './OutdatedPackagesProcessor'
-export * from './OutdatedPullRequestStrategy'
 export * from './PackageUpdater'
 export * from './SucceededResult'
