+++ src/options/cliOptions.ts
@@ -1,5 +1,5 @@
-import { OutdatedPullRequestStrategy } from '../core'
 import { LogLevel } from '../logger'
+import { OutdatedPullRequestStrategy } from '../outdated-package-processor'
 import { PackageManagerName } from '../package-manager'
 import type { CLIOption } from './CLIOption'
 import { OptionType } from './OptionType'
