+++ src/outdated-package-processor/index.ts
@@ -1,0 +1,3 @@
+export * from './OutdatedPackageProcessor'
+export * from './OutdatedPackageProcessorCreator'
+export * from './OutdatedPullRequestStrategy'
